refactor: move client and services to platform package
• Create new platform/ package to organize platform-related modules
• Move client.py to platform/client.py with updated relative imports
• Move services/ directory to platform/services/ maintaining structure
• Update all import paths across bindings and server modules
• Fix test imports to use new platform.services.* paths
• Maintain backward compatibility through proper __init__.py exports
