+++ src/itential_mcp/platform/services/adapters.py
@@ -5,7 +5,7 @@ import asyncio
 
 from itential_mcp.core import exceptions
 
-from itential_mcp.services import ServiceBase
+from itential_mcp.platform.services import ServiceBase
 
 
 class Service(ServiceBase):
