+++ src/itential_mcp/platform/services/configuration_manager.py
@@ -7,7 +7,7 @@ import ipsdk
 
 from itential_mcp.core import exceptions
 
-from itential_mcp.services import ServiceBase
+from itential_mcp.platform.services import ServiceBase
 
 
 class Service(ServiceBase):
