+++ src/itential_mcp/platform/services/health.py
@@ -3,7 +3,7 @@
 
 from typing import Dict, Any
 
-from itential_mcp.services import ServiceBase
+from itential_mcp.platform.services import ServiceBase
 
 
 class Service(ServiceBase):
