+++ src/itential_mcp/platform/services/integrations.py
@@ -5,7 +5,7 @@
 from itential_mcp.core import exceptions
 
 
-from itential_mcp.services import ServiceBase
+from itential_mcp.platform.services import ServiceBase
 
 
 class Service(ServiceBase):
