+++ src/itential_mcp/platform/services/lifecycle_manager.py
@@ -7,7 +7,7 @@ from typing import Mapping, Any, Sequence
 
 from itential_mcp.core import exceptions
 
-from itential_mcp.services import ServiceBase
+from itential_mcp.platform.services import ServiceBase
 
 
 class Service(ServiceBase):
