+++ src/itential_mcp/platform/services/transformations.py
@@ -3,7 +3,7 @@
 
 from typing import Mapping, Any
 
-from itential_mcp.services import ServiceBase
+from itential_mcp.platform.services import ServiceBase
 
 
 class Service(ServiceBase):
