+++ src/itential_mcp/platform/services/workflow_engine.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2025 Itential, Inc
 # GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
-from itential_mcp.services import ServiceBase
+from itential_mcp.platform.services import ServiceBase
 
 
 class Service(ServiceBase):
