+++ tests/test_services_adapters.py
@@ -6,7 +6,7 @@ import asyncio
 from unittest.mock import AsyncMock, MagicMock, patch
 
 from itential_mcp.core import exceptions
-from itential_mcp.services.adapters import Service
+from itential_mcp.platform.services.adapters import Service
 
 
 class TestAdaptersService:
@@ -421,7 +421,7 @@ class TestServiceIntegration:
     @pytest.mark.asyncio
     async def test_service_inherits_from_servicebase(self, service):
         """Test that Service properly inherits from ServiceBase"""
-        from itential_mcp.services import ServiceBase
+        from itential_mcp.platform.services import ServiceBase
 
         assert isinstance(service, ServiceBase)
