+++ tests/test_services_automation_studio.py
@@ -5,7 +5,7 @@ import pytest
 from unittest.mock import AsyncMock, MagicMock
 
 from itential_mcp.core import exceptions
-from itential_mcp.services.automation_studio import Service
+from itential_mcp.platform.services.automation_studio import Service
 
 
 class TestAutomationStudioService:
