+++ tests/test_services_lifecycle_manager.py
@@ -5,7 +5,7 @@ import pytest
 from unittest.mock import AsyncMock, MagicMock, patch
 
 from itential_mcp.core import exceptions
-from itential_mcp.services.lifecycle_manager import Service
+from itential_mcp.platform.services.lifecycle_manager import Service
 from ipsdk.platform import AsyncPlatform
 
 
@@ -31,7 +31,7 @@ class TestService:
 
     def test_service_inheritance(self, service):
         """Test Service inherits from ServiceBase"""
-        from itential_mcp.services import ServiceBase
+        from itential_mcp.platform.services import ServiceBase
 
         assert isinstance(service, ServiceBase)
         assert hasattr(service, "client")
