+++ tests/test_services_transformations.py
@@ -4,7 +4,7 @@
 import pytest
 from unittest.mock import AsyncMock, MagicMock
 
-from itential_mcp.services.transformations import Service
+from itential_mcp.platform.services.transformations import Service
 
 
 class TestTransformationsService:
@@ -272,7 +272,7 @@ class TestServiceIntegration:
     @pytest.mark.asyncio
     async def test_service_inherits_from_servicebase(self, service):
         """Test that Service properly inherits from ServiceBase"""
-        from itential_mcp.services import ServiceBase
+        from itential_mcp.platform.services import ServiceBase
 
         assert isinstance(service, ServiceBase)
